GroupOverviewコンポーネント追加
